refactor: move impl for header to features directory
