+++ app/src/features/header/index.ts
@@ -1,0 +1,1 @@
+export { Header } from "./components/header";
